### 2023-09-28 #5892 @editorialbot
[PRE REVIEW]: ddtlcm: An R package for overcoming weak separation in Bayesian latent class analysis via tree-regularization
**Submitting author:** @limengbinggz (<a href="http://orcid.org/0000-0002-2264-8006">Mengbing Li</a>)
**Repository:** https://github.com/limengbinggz/ddtlcm
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.2
