### 2023-09-28 #5892 @arfon
@limengbinggz – I'm going to query the scope of this submission as 1) it's on the smaller side of what we typically publish in JOSS and 2) the quality seems a little suspect in a few places (e.g., no tests as far as I can see).

This scope review will take a couple of weeks.
