### 2023-10-18 #5892 @limengbinggz
> @editorialbot query scope
>
> @limengbinggz – I'm going to query the scope of this submission as 1) it's on the smaller side of what we typically publish in JOSS and 2) the quality seems a little suspect in a few places (e.g., no tests as far as I can see).
>
> This scope review will take a couple of weeks.

Thank you for pointing the potential issues out.

For 1), I would like to argue that this submission represents substantial scholarly effort. The software, which has been publicly available for less than six months, boasts a meticulously maintained commit history. It comprises well over 1000 lines of R code and adheres to the documentation standards of R CRAN. Methodologically, the software implements a Bayesian latent class model that effectively addresses a long-standing weak separation issue in mixture models. To the best of my knowledge, there are currently no other software packages designed to address this specific statistical challenge. Scientifically, the implemented method is inspired by the need for precise derivation of dietary patterns within the realm of nutritional epidemiology. The related method paper illustrates an application of this software to analyzing dietary behaviors in small subpopulations, a task hitherto unfeasible with existing latent class model software due to issues with class separation. Overall, this submission underscores the innovative and impactful contributions of this software to the field of statistical modeling and nutritional epidemiology.

For 2), tests have been added to the GitHub repo under `tests/`.
