### 2023-12-08 #5892 @limengbinggz
> Hey @limengbinggz thank you for your submission! Can you please check out this page: https://reviewers.joss.theoj.org/lookup and suggest a few more reviewers? 😄

Thank you for the follow up. Here are some reviewers potentially relevant to keywords `R` and `Bayesian`:

- jamesmbaazam
- nhtran93
- jamesuanhoro
- larryshamalama
- EugeneHao
