### 2023-12-11 #5892 @Nikoleta-v3
👋🏻 @nhtran93 @jamesuanhoro @larryshamalama would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "ddtlcm: An R package for overcoming weak separation in Bayesian latent class analysis via tree-regularization". You can find more information at the top of this Github issue 🆙
