### 2024-01-10 #5892 @jamesuanhoro
> 👋🏻 @nhtran93 @jamesuanhoro @larryshamalama would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "ddtlcm: An R package for overcoming weak separation in Bayesian latent class analysis via tree-regularization". You can find more information at the top of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!

You can sign me up to review.
