### 2024-02-21 limengbinggz/ddtlcm#6 @larryshamalama
JOSS Review: Community guidelines
> **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

For JOSS [#6620](https://github.com/openjournals/joss-reviews/issues/6220), they seem to require some guidelines on reporting issues or contributing to the software. For bigger projects, some contributing guidelines can be beneficial (e.g. [one that I personally am actively involved in](https://github.com/pymc-devs/pymc/blob/main/CONTRIBUTING.md)). Perhaps one or two sentences in a section such as the following can be added in the README.md file:

### Contributing (or Getting help?)

Please report bugs by opening an [issue](https://github.com/limengbinggz/ddtlcm/issues/new). If you wish to contribute, please refer to... If you have questions, you can open a discussion thread (<- you need to enable this for this repo IIRC).
