### 2024-02-21 #6220 @larryshamalama
Hi @limengbinggz, great software and great method! From one fellow biostats PhD to another, congratulations on your hard work :)

Below are my questions/comments, on top of some other quick things that I am opening as issue(s)/PR(s) in your repo

### 1. Weak Separation and State of the Field

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

It took me a while to understand what is meant by "weak separation", which is the focal point of this work. Perhaps this is because I am not so familiar with Bayesian tree-based methods... I was initially not sure if using other software that you mentioned (e.g. `poLCA`, `BayesLCA`, `randomLCA`) would yield worse results on your simulated data but Figure 3 in your methods paper seems to confirm this. I am thinking that it would be beneficial to clarify this result here since it would motivate further why we should use your package and not the other ones. What do you think?

**Minor comment**: lines 21, 22: "classes that share proximity to one another in the tree are shrunk towards ancestral classes _a priori_" Do you think that you can massage this a bit? I'm not sure if I fully understand, but this sentence seems importance since it highlights on a higher level what this method is doing under the food (re: **summary** bullet point above).

### 2. 50 burn-in, 100 posterior draws

In your example, you seem to use 50 burn-in draws and 100 posterior draws. Is that sufficient? As a user, how would I know when the MCMC converges with your software? If I am thinking of conventional MCMC-ing, these seem like low numbers, especially that, in your example, you use $N = 496$ which is larger than the total number of draws. In section 5.1 of your methods paper, you seem to use much more draws in the Gibbs sampler (7,000 and 12,000).

### 3. Singleton node warnings

I am getting many `In checkTree(object) : Tree contains singleton nodes.` when running your quickstart. I imagine that this means that, for a specific category, there is only one member, which can happen, but just wanted to double check that this can be normal/okay. Can you briefly comment on this?
