### 2024-02-26 571366b
Safer code using sequence functions
- Converted `1:length(x)` code to `seq_along(x)`
- Converted `1:x` code to `seq_len(x)`
- Using `&&` and `||` in if statements since if statements are scalar conditionals
