### 2024-02-26 limengbinggz/ddtlcm#10 @jamesuanhoro
Trying to assess some elements within the object returned by ddtlcm_fit() results in an error
Reproducible code:

```r
library(ddtlcm)
# load the data
data(parameter_diet)
# unlist the elements into variables in the global environment
list2env(setNames(parameter_diet, names(parameter_diet)), envir = globalenv())
# look at items in group 1
g <- 1
# indices of the items in group 1
item_membership_list[g]

# The name of the list element is the major food group label
item_name_list[g]

# number of individuals
N <- 496
# random seed to generate node parameters given the tree
seed_parameter <- 1
# random seed to generate multivariate binary observations from LCM
seed_response <- 1
# simulate data given the parameters
sim_data <- simulate_lcm_given_tree(tree_phylo, N,
  class_probability, item_membership_list, Sigma_by_group,
  root_node_location = 0, seed_parameter = 1, seed_response = 1
)

set.seed(999)
# number of latent classes, same as number of leaves on the tree
K <- 6
result_diet <- ddtlcm_fit(
  K = K, data = sim_data$response_matrix,
  item_membership_list = item_membership_list, total_iters = 100
)
print(result_diet)

result_diet$tree_samples$tree_list
# [[1]]
# Error in checkPhylo4Data(object) :
#   The row names of tree data do not match the node numbers
# In addition: Warning message:
# In checkTree(object) : Tree contains singleton nodes.
```

```r
sessionInfo()
# R version 4.3.2 (2023-10-31)
# Platform: x86_64-pc-linux-gnu (64-bit)
# Running under: Pop!_OS 20.04 LTS

# Matrix products: default
# BLAS:   /usr/lib/x86_64-linux-gnu/atlas/libblas.so.3.10.3
# LAPACK: /usr/lib/x86_64-linux-gnu/atlas/liblapack.so.3.10.3;  LAPACK version 3.9.0

# locale:
#  [1] LC_CTYPE=en_GB.UTF-8       LC_NUMERIC=C               LC_TIME=en_GB.UTF-8        LC_COLLATE=en_GB.UTF-8     LC_MONETARY=en_GB.UTF-8    LC_MESSAGES=en_GB.UTF-8
#  [7] LC_PAPER=en_GB.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=en_GB.UTF-8 LC_IDENTIFICATION=C

# time zone: America/Chicago
# tzcode source: system (glibc)

# attached base packages:
# [1] stats     graphics  grDevices utils     datasets  methods   base

# other attached packages:
# [1] ddtlcm_0.1.1

# loaded via a namespace (and not attached):
#  [1] ade4_1.7-22          tidyselect_1.2.0     dplyr_1.1.4          R.utils_2.12.3       fastmap_1.1.1        lazyeval_0.2.2       combinat_0.0-8       XML_3.99-0.16.1
#  [9] digest_0.6.34        lifecycle_1.0.4      tidytree_0.4.6       magrittr_2.0.3       compiler_4.3.2       rlang_1.1.3          progress_1.2.3       tools_4.3.2
# [17] utf8_1.2.4           data.table_1.15.0    ggsignif_0.6.4       prettyunits_1.2.0    label.switching_1.8  scatterplot3d_0.3-44 plyr_1.8.9           xml2_1.3.6
# [25] RNeXML_2.4.11        aplot_0.2.2          abind_1.4-5          R.cache_0.16.0       purrr_1.0.2          R.oo_1.26.0          grid_4.3.2           fansi_1.0.6
# [33] ggpubr_0.6.0         colorspace_2.1-0     ggplot2_3.5.0        scales_1.3.0         MASS_7.3-60.0.1      cli_3.6.2            crayon_1.5.2         treeio_1.26.0
# [41] generics_0.1.3       rstudioapi_0.14      ggtree_3.10.1        httr_1.4.7           reshape2_1.4.4       ape_5.7-1            cachem_1.0.8         stringr_1.5.1
# [49] parallel_4.3.2       ggplotify_0.1.2      matrixStats_1.2.0    vctrs_0.6.5          yulab.utils_0.1.4    Matrix_1.6-5         jsonlite_1.8.8       carData_3.0-5
# [57] car_3.1-2            gridGraphics_0.5-1   hms_1.1.3            patchwork_1.2.0      BayesLogit_2.1       rstatix_0.7.2        phylobase_0.8.12     tidyr_1.3.1
# [65] poLCA_1.6.0.1        glue_1.7.0           ggtext_0.1.2         stringi_1.8.3        gtable_0.3.4         munsell_0.5.0        tibble_3.2.1         styler_1.9.1
# [73] pillar_1.9.0         truncnorm_1.0-9      R6_2.5.1             Rdpack_2.6           lpSolve_5.6.20       lattice_0.22-5       extraDistr_1.10.0    rbibutils_2.2.16
# [81] R.methodsS3_1.8.2    backports_1.4.1      gridtext_0.1.5       memoise_2.0.1        broom_1.0.5          ggfun_0.1.4          Rcpp_1.0.12          rncl_0.8.7
# [89] uuid_1.2-0           nlme_3.1-163         fs_1.6.3             pkgconfig_2.0.3
```

Link: https://github.com/openjournals/joss-reviews/issues/6220
