### 2024-02-26 limengbinggz/ddtlcm#11 @jamesuanhoro
How to assess parameter convergence?
It would be really helpful for users to know how to assess parameter convergence.

Using the example in [the paper](/limengbinggz/ddtlcm/blob/main/JOSS/paper.pdf), it seems something like this works(?)

```r
# incorrect: does not discard burnin
posterior_samples <- t(result_diet$class_probs_samples)
posterior_draws <- posterior::as_draws(posterior_samples)
posterior::summarise_draws(posterior_draws)
bayesplot::mcmc_trace(posterior_draws)
```

It's nice that the content of the results object is described in the manual -- that helped me guess this example. It would be helpful for a reader to see a cleaner version of the above somewhere e.g. in a vignette since convergence assessment is a key part of MCMC.

openjournals/joss-reviews/issues/6220
