### 2024-02-26 limengbinggz/ddtlcm#12 @jamesuanhoro
Assessing the calibration of MCMC implementation
Since you wrote the MCMC sampling algorithms, it would be good to have some guarantees that the program produces calibrated Bayesian inference.

To my knowledge, the best way to assess this is [simulation based calibration](https://arxiv.org/abs/1804.06788), see recent developments in: https://hyunjimoon.github.io/SBC/

So my recommendation is to conduct an SBC study possibly using a smaller version of the example study for speed.

Open to discussion about this.

openjournals/joss-reviews/issues/6220
