### 2024-02-26 8d44cc2
A few changes/suggestions:
- If number of iterations exceeds 10, update user at every 10% -- approach is not exactly elegant
- Show percentages in treemap
- rea-valued => real-valued in documentation
