### 2024-03-14 0915946
address incorrect dimension when accessing data with NA rows from a phylo4d object
