+++ vignettes/ddtlcm-demo.Rmd
@@ -34,6 +34,7 @@ knit_hooks$set(output = function(x, options) {
   }
   hook_output(x, options)
 })
+options(warn = -1)
 ```
 
 
@@ -144,18 +145,20 @@ On the left is the MAP tree estimated from the real HCHS/SOL data, with leaf nod
 
 ## Apply DDT-LCM
 
-We assume that the number of latent classes $K = 6$ is known. Throughout the paper, we use divergence function $a(t) = c / (1-t)$ to parameterize the DDT branching process, where $c > 0$ is a hyperparameter. To use the function `ddtlcm_fit`, we need to specify the number of classes (`K`), a matrix of multivariate binary observations (`data`), a list of item group memberships (`item_membership_list`), and the number of posterior samples to collect (`total_iters`). In Section 5.1 of the manuscript, we set `total_iters = 15000`. For the purpose of quickly illustrating the output, here we run the function with a smaller number `total_iters = 100`.
+We assume that the number of latent classes $K = 6$ is known. Throughout the paper, we use divergence function $a(t) = c / (1-t)$ to parameterize the DDT branching process, where $c > 0$ is a hyperparameter. To use the function `ddtlcm_fit`, we need to specify the number of classes (`K`), a matrix of multivariate binary observations (`data`), a list of item group memberships (`item_membership_list`), and the number of posterior samples to collect (`total_iters`). In Section 5.1 of the manuscript, we set `total_iters = 15000`. For the purpose of quickly illustrating the output, here we run the function with a smaller number `total_iters = 100`. Note that ``suppressWarnings`` is used to prevent the warning "Tree contains singleton nodes" message from the phylobase package.
 
 ```{r}
 set.seed(999)
 # number of latent classes, or number of leaves on the tree
 K <- 6
+suppressWarnings({ 
 system.time({
   result <- ddtlcm_fit(K = K, data = response_matrix, 
                      item_membership_list = item_membership_list, total_iters = 100)
 })
-
+})
 print(result)
+# result$tree_samples$tree_list
 ```
 
 We next summarize the posterior chain by discarding the first 50 iterations as burn-ins (`burnin = 50`). To deal with identifiability of finite mixture models, we perform post-hoc label switching using the Equivalence Classes Representatives (ECR) method by specifying `relabel = T`. To save space in the document, we do not print the summary result here (`be_quiet = T`).
