### 2024-03-14 limengbinggz/ddtlcm#10 @limengbinggz
Thank you for pointing out the issue. We have addressed this in commit [0915946](https://github.com/limengbinggz/ddtlcm/commit/0915946b6f31246f5f18018981d3b2a59b41aa0c). The issue arises from the ``sample_tree_topology`` function. When when accessing data associated with a tree of class "phylo4d", rows with all NA's in the data are dropped automatically, resulting in incorrect dimension of the dataset, which is supposed to have 2K (twice the number of leaf nodes or latent classes) rows and J (number of items) columns. The rows associated with the internal nodes of the tree are coded as NA's because we do not care about their exact values.
