### 2024-03-14 ac5b83d
In README: Add a note about warning; add a Contributing section
