+++ README.md
@@ -63,9 +63,15 @@ plot(dep)
 Examples 
 ---------
 
-* A simple workflow using semi-synthetic data is provided in ![](inst/ddtlcm_workflow_example.pdf)
+* A simple [workflow](inst/ddtlcm_workflow_example.pdf) using semi-synthetic data is provided.
 
 * *ddtlcm* estimates the tree over classes and class profiles simultaneously ![](inst/ddtlcm_output_example.png)
+
+
+
+Contributing (or Getting help?)
+---------
+Please report bugs by opening an [issue](https://github.com/limengbinggz/ddtlcm/issues/new). If you wish to contribute, please make a pull request. If you have questions, you can open a [discussion thread](https://github.com/limengbinggz/ddtlcm/discussions).
 
 
 
@@ -74,7 +80,9 @@ Note
 
 * When running some functions in the package, such as ``ddtlcm_fit``, a warning that "Tree contains singleton nodes" may be displayed. This warning originates from the checkPhylo4 in the phylobase package to perform basic checks on the validity of S4 phylogenetic objects. We would like to point out that seeing such warnings shall not pose any concerns about the statistical validity of the implemented algorithm. This is because any tree generaetd from a DDT process contains a singleton node (having only one child node) as the root node. To avoid repeated appearances of this warning, we recommend either of the followings:
 
-    - 
+    - Wrapping around the code via ``suppressWarnings({ code_that_will_generate_singleton_warning })``;
+
+    - Setting ``options(warn = -1)`` globally. This may be dangerous because other meaningful warnings may be ignored.
 
 
 
