### 2024-03-14 limengbinggz/ddtlcm#6 @limengbinggz
Thank you for point out. This is fixed in commit [df7200a](https://github.com/limengbinggz/ddtlcm/commit/df7200a7940988e455aa75129008af6b93f2beee).
