+++ JOSS/references.bib
@@ -59,14 +59,19 @@
   year={2023}
 }
 
-@article{neal2003density,
-  title={Density modeling and clustering using {Dirichlet} diffusion trees},
-  author={Neal, Radford M},
-  journal={Bayesian Statistics},
-  volume={7},
-  pages={619--629},
-  year={2003}
+@incollection{neal2003density,
+    author = {Neal, Radford M},
+    isbn = {9780198526155},
+    title = "{Density Modeling and Clustering Using Dirichlet Diffusion Trees}",
+    booktitle = "{Bayesian Statistics 7: Proceedings of the Seventh Valencia International Meeting}",
+    publisher = {Oxford University Press},
+    year = {2003},
+    month = {07},
+    doi = {10.1093/oso/9780198526155.003.0042},
+    url = {https://doi.org/10.1093/oso/9780198526155.003.0042},
+    eprint = {https://academic.oup.com/book/0/chapter/422208359/chapter-pdf/52446748/isbn-9780198526155-book-part-42.pdf},
 }
+
 
 @article{papastamoulis2014HandlingLabel,
   title = {Handling the Label Switching Problem in Latent Class Models Via the {ECR} Algorithm},
