### 2024-03-15 #6220 @limengbinggz
> @limengbinggz, could you please add the DOI for one of your references? See: [#6220 (comment)](https://github.com/openjournals/joss-reviews/issues/6220#issuecomment-1888851682)

Thank you for pointing out. We have added the DOI for the reference. The paper is updated in commit
[4f9a157](https://github.com/limengbinggz/ddtlcm/commit/4f9a1571994d2b949bc2178d5d8c6dcba8037982)
