+++ man/logllk_location.Rd
@@ -34,8 +34,8 @@ Compute the marginal log likelihood of the parameters on the leaves of a tree
 }
 \seealso{
 Other likelihood functions: 
-\code{\link{logllk_ddt_lcm}()},
 \code{\link{logllk_ddt}()},
+\code{\link{logllk_ddt_lcm}()},
 \code{\link{logllk_div_time_one}()},
 \code{\link{logllk_div_time_two}()},
 \code{\link{logllk_lcm}()},
