### 2024-03-25 #6220 @limengbinggz
> Hi @limengbinggz, great software and great method! From one fellow biostats PhD to another, congratulations on your hard work :)
>
> Below are my questions/comments, on top of some other quick things that I am opening as issue(s)/PR(s) in your repo
>
> ### 1. Weak Separation and State of the Field
> > State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> It took me a while to understand what is meant by "weak separation", which is the focal point of this work. Perhaps this is because I am not so familiar with Bayesian tree-based methods... I was initially not sure if using other software that you mentioned (e.g. `poLCA`, `BayesLCA`, `randomLCA`) would yield worse results on your simulated data but Figure 3 in your methods paper seems to confirm this. I am thinking that it would be beneficial to clarify this result here since it would motivate further why we should use your package and not the other ones. What do you think?
>
> **Minor comment**: lines 21, 22: "classes that share proximity to one another in the tree are shrunk towards ancestral classes _a priori_" Do you think that you can massage this a bit? I'm not sure if I fully understand, but this sentence seems importance since it highlights on a higher level what this method is doing under the food (re: **summary** bullet point above).
>
> ### 2. 50 burn-in, 100 posterior draws
> In your example, you seem to use 50 burn-in draws and 100 posterior draws. Is that sufficient? As a user, how would I know when the MCMC converges with your software? If I am thinking of conventional MCMC-ing, these seem like low numbers, especially that, in your example, you use N=496 which is larger than the total number of draws. In section 5.1 of your methods paper, you seem to use much more draws in the Gibbs sampler (7,000 and 12,000).
>
> ### 3. Singleton node warnings
> I am getting many `In checkTree(object) : Tree contains singleton nodes.` when running your quickstart. I imagine that this means that, for a specific category, there is only one member, which can happen, but just wanted to double check that this can be normal/okay. Can you briefly comment on this?

Thank you very much for reviewing the software as well as the methods paper!

1. Thanks for pointing out the problem in writing. As you suggested, we have added a clarifying sentence to specifically point out why the existing packages fail, by referring to the results in our method paper: "These phenomena have been demonstrated using ``poLCA`` and ``BayesLCA``, which are more relevant to our stated problem of weak class separation, in Figures 2 and 3 of @li2023tree."

As for the sentence regarding how classes are shrunk, we have clarified this point by changing the original sentence into "classes that are closer to one another in the binary tree are encouraged to share more similar profiles, and their profiles are shrunk towards their common ancestral classes \textit{a priori}, with the degree of shrinkage varying across pre-specified item groups defined thematically with clinical significance. "

2. I definitely agree that 100 posterior draws are far from sufficient. However, the sampling algorithm implemented in this package is quite computationally intensive. For the sake of time and for the purpose of brief illustration, we want to demonstrate the functionality of the sampling function ``ddtlcm_fit``. But to your good point that to provide a comprehensive view of a full posterior chain, we have added a pre-saved dataset named "result_diet_fullchain" with 8000 posterior samples, to accompany the package. In the paper, we have added a sentence to clarify this point in the last paragraph of section "Model Fitting": "To have a more comprehensive view of of the results obtained from 8000 posterior draws, we can load data named `result_diet_fullchain` to perform posterior summaries as described by the steps in the following sections."

3. Thank you for pointing this out. In fact, the trees we are dealing with in our model all have singleton nodes by design. This warning message originates from the ``checkPhylo4`` function in the phylobase package to perform basic checks on the validity of S4 phylogenetic objects, where phylogenetic trees usually want to avoid singleton nodes. Therefore, this warning shall not be concerning. To clarify this point, we have added the following Note section in README of the package:

* When running some functions in the package, such as ``ddtlcm_fit``, a warning that "Tree contains singleton nodes" may be displayed. This warning originates from the ``checkPhylo4`` function in the phylobase package to perform basic checks on the validity of S4 phylogenetic objects. We would like to point out that seeing such warnings shall not pose any concerns about the statistical validity of the implemented algorithm. This is because any tree generaetd from a DDT process contains a singleton node (having only one child node) as the root node. To avoid repeated appearances of this warning, we recommend either of the followings:

    - Wrapping around the code via ``suppressWarnings({ code_that_will_generate_singleton_warning })``;

    - Setting ``options(warn = -1)`` globally. This may be dangerous because other meaningful warnings may be ignored.
