### 2024-03-25 b7f1f2a
create a data file with smaller size
