+++ R/RcppExports.R
@@ -1,15 +1,0 @@
-# Generated by using Rcpp::compileAttributes() -> do not edit by hand
-# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-expit <- function(x) {
-    .Call('_ddtlcm_expit', PACKAGE = 'ddtlcm', x)
-}
-
-tr <- function(x) {
-    .Call('_ddtlcm_tr', PACKAGE = 'ddtlcm', x)
-}
-
-riwishart <- function(df, S) {
-    .Call('_ddtlcm_riwishart', PACKAGE = 'ddtlcm', df, S)
-}
-
