+++ R/plot_functions.R
@@ -3,8 +3,8 @@
 #' @param plot_option option to select which part of the plot to return. If "all", return
 #'  the plot of MAP tree on the left and the plot of class profiles on the right. If "profile",
 #'  only return the plot of class profiles. If "tree", only return the plot of MAP tree.
-#' @param item_name_list a named list of G elements, where the g-th element contains a vector 
-#'  of item names for items in `item_membership_list[[g]]`. The name of the g-th element is 
+#' @param item_name_list a named list of G elements, where the g-th element contains a vector
+#'  of item names for items in `item_membership_list[[g]]`. The name of the g-th element is
 #'  the name of the major item group.
 #' @param color_palette a vector of color names. Default is a color-blinded friendly palette.
 #' @param log Default argument passed to plot(). Not used.
@@ -12,20 +12,20 @@
 #' @method plot summary.ddt_lcm
 #' @importFrom ggplot2 geom_errorbar
 #' @importFrom ggpubr ggarrange
-#' @return a ggplot2 object. If plot_option is "all", then a plot with maximum a posterior 
-#'  tree structure on the left and a bar plot of item response probabilities (with 95% 
-#'  credible intervals and class probabilities) on the right. If plot_option is 
-#'  "profile", then only a bar plot of item response probabilities. If plot_option is 
+#' @return a ggplot2 object. If plot_option is "all", then a plot with maximum a posterior
+#'  tree structure on the left and a bar plot of item response probabilities (with 95%
+#'  credible intervals and class probabilities) on the right. If plot_option is
+#'  "profile", then only a bar plot of item response probabilities. If plot_option is
 #'  "tree", then only a plot of the tree structure.
-#' @export 
+#' @export
 #' @examples
-#' data(result_diet)
-#' burnin <- 50
-#' summarized_result <- summary(result_diet, burnin, relabel = TRUE, be_quiet = TRUE)
+#' data(result_diet_1000iters)
+#' burnin <- 500
+#' summarized_result <- summary(result_diet_1000iters, burnin, relabel = TRUE, be_quiet = TRUE)
 #' plot(x = summarized_result, item_name_list = NULL, plot_option = "all")
-plot.summary.ddt_lcm <- function(x, log=TRUE, 
+plot.summary.ddt_lcm <- function(x, log=TRUE,
                                  plot_option = c("all", "profile", "tree"),
-                                 item_name_list = NULL, color_palette = c("#E69F00", "#56B4E9", "#009E73", "#000000", 
+                                 item_name_list = NULL, color_palette = c("#E69F00", "#56B4E9", "#009E73", "#000000",
                                                                           "#0072B2", "#D55E00", "#CC79A7", "#F0E442", "#999999"),
                                  ...){
   plot_option = match.arg(plot_option, c("all", "profile", "tree"))
@@ -36,8 +36,8 @@ plot.summary.ddt_lcm <- function(x, log=TRUE,
   class_probability <- x$class_probs_summary[,"Mean"]
   class_probability_lower <- x$class_probs_summary[,"2.5%"]
   class_probability_higher <- x$class_probs_summary[,"97.5%"]
-  plots <- plot_tree_with_barplot(tree_with_parameter, response_prob, item_membership_list, 
-                         item_name_list, class_probability, class_probability_lower, class_probability_higher, 
+  plots <- plot_tree_with_barplot(tree_with_parameter, response_prob, item_membership_list,
+                         item_name_list, class_probability, class_probability_lower, class_probability_higher,
                          color_palette, return_separate_plots = TRUE)
[…]
