+++ R/predict.R
@@ -1,6 +1,6 @@
 #' Prediction of class memberships from posterior summaries
 #' @description Predict individual class memberships based on posterior summary (point estimates
-#'  of model parameters). The predicted class memberships are modal assignments. 
+#'  of model parameters). The predicted class memberships are modal assignments.
 #' @param object a "summary.ddt_lcm" object
 #' @param data an NxJ matrix of multivariate binary responses, where
 #'   N is the number of individuals, and J is the number of granular items
@@ -14,11 +14,11 @@
 #'  is the probability that the i-th individual is predicted to belong to class k.}
 #' }
 #' @examples
-#' data(result_diet)
-#' burnin <- 50
-#' summarized_result <- summary(result_diet, burnin, relabel = TRUE, be_quiet = TRUE)
-#' predicted <- predict(summarized_result, result_diet$data)
-predict.summary.ddt_lcm <- function(object, data, ...){ 
+#' data(result_diet_1000iters)
+#' burnin <- 500
+#' summarized_result <- summary(result_diet_1000iters, burnin, relabel = TRUE, be_quiet = TRUE)
+#' predicted <- predict(result_diet_1000iters, result_diet$data)
+predict.summary.ddt_lcm <- function(object, data, ...){
   class_probability <- object$class_probs_summary[,'Mean']
   K <- length(class_probability)
   N <- nrow(data)
@@ -31,7 +31,7 @@ predict.summary.ddt_lcm <- function(object, data, ...){
   class_assignments <- rep(0, N)
   # for each individual
   for (i in 1:N) {
-    class_assignments[i] <- sample(1:K, 1, prob = predictive_probs[,i]) 
+    class_assignments[i] <- sample(1:K, 1, prob = predictive_probs[,i])
   }
   return(list(class_assignments = class_assignments, predictive_probs = t(predictive_probs)))
 }
@@ -55,14 +55,14 @@ predict.summary.ddt_lcm <- function(object, data, ...){
 #'  is the probability that the i-th individual is predicted to belong to class k.}
 #' }
 #' @examples
-#' data(result_diet)
-#' burnin <- 50
-#' predicted <- predict(result_diet, result_diet$data, burnin)
-predict.ddt_lcm <- function(object, data, burnin = 3000, ...){ 
+#' data(result_diet_1000iters)
+#' burnin <- 500
+#' predicted <- predict(result_diet_1000iters, result_diet_1000iters$data, burnin)
+predict.ddt_lcm <- function(object, data, burnin = 3000, ...){
   total_iters <- length(object$loglikelihood)
   K <- object$setting$K
   N <- nrow(data)
-  
+
   class_assignment_freq <- matrix(0, nrow = N, ncol = K)
   for (iter in (burnin+1):total_iters) {
     class_probability <- object$class_probs_samples[,iter]
