+++ R/data.R
@@ -29,28 +29,15 @@
 #' Result of fitting DDT-LCM to a semi-synthetic data example
 #'
 #' This is a "ddtlcm" object obtained from running `ddtlcm_fit` to a semi-synthetic dataset with
-#'  100 posterior samples (for the sake of time). See \code{\link{ddtlcm_fit}} for description of 
+#'  1000 posterior samples (for the sake of time). See \code{\link{ddtlcm_fit}} for description of 
 #'  the object.
 #'
 #' @docType data
 #' @keywords datasets
-#' @name result_diet_shortchain
-#' @usage data(result_diet_shortchain)
+#' @name result_diet_1000iters
+#' @usage data(result_diet_1000iters)
 #' @format A list with 8 elements
-"result_diet_shortchain"
-
-#' Result of fitting DDT-LCM to a semi-synthetic data example
-#'
-#' This is a "ddtlcm" object obtained from running `ddtlcm_fit` to a semi-synthetic dataset with
-#'  8000 posterior samples (for the sake of time). See \code{\link{ddtlcm_fit}} for description of 
-#'  the object.
-#'
-#' @docType data
-#' @keywords datasets
-#' @name result_diet_fullchain
-#' @usage data(result_diet_fullchain)
-#' @format A list with 8 elements
-"result_diet_fullchain"
+"result_diet_1000iters"
 
 
 #' Synthetic data example

[…]
