### 2024-03-25 cf80cb9
add a plot function for trace plots
