+++ R/plot_functions.R
@@ -1,3 +1,105 @@
+#' Create trace plots of DDT-LCM parameters
+#' @param x a "ddt_lcm" object
+#' @param parameter_names a character vector to specify the parameters to be plotted. Each element can take the
+#'  be 1) of format "parameter_index" to plot specific parameters with specific indices, 2) of format "parameter"
+#'  to plot the parameters across all indices, or 3) equal to "all" to plot all parameters in the model. For 1),
+#'  the item response probabilities should be named "responseprob_class,group,item"; the class probabilities should be
+#'  named "classprob_class"; the divergence function parameter is "c"; the group-specific diffusion variances
+#'  should be named "diffusionvar_group". For 2), "responseprob" to plot all item response probabilities;
+#'  "classprob" to plot all class probabilities; "diffusionvar" to plot all diffusion variances.
+#' @param burnin the number of posterior samples as burn-in, which will not be plotted.
+#' @param \dots	Further arguments passed to each method
+#' @method plot ddt_lcm
+#' @return NULLs
+#' @export
+#' @examples
+#' data(result_diet_1000iters)
+#' # we will plot: "responseprob_1,1,1" for the class 1 response probability of item 3 in major group 2;
+#' #   "classprob_1" for the probability of being assigned to class 1; "c" for the divergence function parameter;
+#' #   and "diffusionvar_1" for the diffusion variance of group 1
+#' plot(x = result_diet_1000iters, c("responseprob_1,1,1", "classprob_1", "c", "diffusionvar_1"), burnin = 500)
+#' # plot all class probabilities
+#' plot(x = result_diet_1000iters, "classprob", burnin = 500)
+#' # plot all diffusion variances
+#' plot(x = result_diet_1000iters, "diffusionvar", burnin = 500)
+plot.ddt_lcm <- function(x, parameter_names = c("responseprob_1,1,1", "classprob_1", "c", "diffusionvar_1"), burnin = 50, ...){
+  # number of classes
+  K <- x$setting$K
+  # number of major item groups
+  G <- x$setting$G
+  # extract list of item indices according to groups
+  item_membership_list <- x$setting$item_membership_list
+  # # get the number of items each group
+  # num_items_per_group <- unlist(lapply(item_membership_list, length))
+  # # cumulative index
+  # cum_num_items_per_group <- c(0, cumsum(num_items_per_group))
+  # number of iterations in total
+  total_iters <- length(x$loglikelihood)
+
+  ## input checking
+  good_names <- grepl("^responseprob|^classprob|^c$|^diffusionvar", parameter_names)
+  bad_names <- which(!good_names)
+  if (length(bad_names) >= 1){
+    stop(paste0("parameter_names[", bad_names, "] invalid. Please check the appropriate format in the function description."))
+  }
+  if (as.integer(burnin) != burnin | burnin < 0 | burnin >= total_iters){
+    stop("The number of burn-in posterior samples, burnin, should be a positive integer smaller than the total number of samples.")
+  }
+
+  if (length(parameter_names) == 1 & parameter_names[1] == "all"){
+    parameter_names <- c("responseprob", "classprob", "c", "diffusionvar")
+  }
+  parameter_names_all <- c()
+  for (param in parameter_names) {
+    param_split <- strsplit(param, "_")[[1]]
+    # get which parameter of the model we are looking at
+    param_type <- param_split[1]
+    # if no indices are specified, then plot all parameters of that type
+    if (length(param_split) == 1 & param_type != "c"){
+      if (param_split == "responseprob") {
+        index_group_item <- unlist(lapply(1:length(item_membership_list), function(x) paste0(x, ",", 1:length(item_membership_list[[x]]))))
[…]
