+++ R/initialization.R
@@ -95,7 +95,7 @@ initialize_hclust <- function(leaf_data, c, c_order=1, method_dist = "euclidean"
   # add root node
   hclust_phylo <- add_root(hclust_phylo, root_edge_length = root_edge_length,
            root_label = "u1", leaf_label = "u2")
-  hclust_phylo <- phylo4d(hclust_phylo)
+  hclust_phylo <- suppressWarnings({ phylo4d(hclust_phylo) })
   # add node labels
   tipLabels(hclust_phylo) <- paste0("v", 1:K)
   nodeLabels(hclust_phylo) <- paste0("u", 1:K)
@@ -116,7 +116,7 @@ initialize_hclust <- function(leaf_data, c, c_order=1, method_dist = "euclidean"
 #'@param c hyparameter of divergence function a(t)
 #'@param c_order equals 1 (default) or 2 to choose divergence function
 #'  a(t) = c/(1-t) or c/(1-t)^2.
-#'@param method_lcm a character. If `random` (default), the initial LCM parameters will be random values. 
+#'@param method_lcm a character. If `random` (default), the initial LCM parameters will be random values.
 #'  If `poLCA`, the initial LCM parameters will be EM algorithm estimates from the \code{poLCA} function.
 #'@param method_dist string specifying the distance measure to be used in dist().
 #'    This must be one of "euclidean" (defaults), "maximum", "manhattan", "canberra", "binary" or "minkowski".
@@ -131,16 +131,16 @@ initialize_hclust <- function(leaf_data, c, c_order=1, method_dist = "euclidean"
 #'    the DDT process with c = 100)
 #'@param fixed_initials a named list of fixed initial values, including
 #' the initial values for tree ("phylo4d"), response_prob, class_probability, class_assignments,
-#' Sigma_by_group, and c. Default is NULL. See 
-#'@param fixed_priors a named list of fixed prior hyperparameters, including the 
+#' Sigma_by_group, and c. Default is NULL. See
+#'@param fixed_priors a named list of fixed prior hyperparameters, including the
 #' the Gamma prior for `c`, inverse-Gamma prior for `sigma_g^2`, and Dirichlet prior
-#' for `pi`. Moreover, we allow for a type III generalized logistic distribution such 
+#' for `pi`. Moreover, we allow for a type III generalized logistic distribution such
 #' that f(`eta`; a_pg) = `theta`. This becomes a standard logistic distribution when a_pg = 1. See
-#' Dalla Valle, L., Leisen, F., Rossini, L., & Zhu, W. (2021). A Pólya–Gamma sampler for a 
+#' Dalla Valle, L., Leisen, F., Rossini, L., & Zhu, W. (2021). A Pólya–Gamma sampler for a
 #' generalized logistic regression. Journal of Statistical Computation and Simulation, 91(14), 2899-2916.
-#' An example input list is 
+#' An example input list is
 #' `list("shape_c" = 1, "rate_c" = 1, "shape_sigma" = rep(2, G), "rate_sigma" = rep(2, G), "a_pg" = 1.0)`, where
-#' `G` is the number of major item groups. Default is NULL. 
+#' `G` is the number of major item groups. Default is NULL.
 #'@param alpha,theta hyparameter of branching probability a(t) Gamma(m-alpha) / Gamma(m+1+theta)
 #'    For DDT, alpha = theta = 0
 #'@param ... optional arguments for the poLCA function
@@ -152,17 +152,17 @@ initialize_hclust <- function(leaf_data, c, c_order=1, method_dist = "euclidean"
 #' # load the MAP tree structure obtained from the real HCHS/SOL data
 #' data(data_synthetic)
 #' # extract elements into the global environment
[…]
