+++ man/predict.summary.ddt_lcm.Rd
@@ -31,5 +31,5 @@ of model parameters). The predicted class memberships are modal assignments.
 data(result_diet_1000iters)
 burnin <- 500
 summarized_result <- summary(result_diet_1000iters, burnin, relabel = TRUE, be_quiet = TRUE)
-predicted <- predict(result_diet_1000iters, result_diet$data)
+predicted <- predict(summarized_result, result_diet_1000iters$data)
 }

[…]
