+++ man/plot.ddt_lcm.Rd
@@ -37,9 +37,9 @@ data(result_diet_1000iters)
 # we will plot: "responseprob_1,1,1" for the class 1 response probability of item 3 in major group 2;
 #   "classprob_1" for the probability of being assigned to class 1; "c" for the divergence function parameter;
 #   and "diffusionvar_1" for the diffusion variance of group 1
-plot(x = result_diet_1000iters, c("responseprob_1,1,1", "classprob_1", "c", "diffusionvar_1"), burnin = 500)
+plot(x = result_diet_1000iters, parameter_names = c("responseprob_1,1,1", "classprob_1", "c", "diffusionvar_1"), burnin = 500)
 # plot all class probabilities
-plot(x = result_diet_1000iters, "classprob", burnin = 500)
+plot(x = result_diet_1000iters, parameter_names = "classprob", burnin = 500)
 # plot all diffusion variances
 plot(x = result_diet_1000iters, "diffusionvar", burnin = 500)
 }
