### 2024-03-26 26ad586
closes #11; correct plot margin error
