+++ NAMESPACE
@@ -49,6 +49,7 @@ importFrom(ggtext,element_markdown)
 importFrom(ggtree,geom_nodelab)
 importFrom(ggtree,geom_tiplab)
 importFrom(ggtree,ggtree)
+importFrom(graphics,par)
 importFrom(label.switching,label.switching)
 importFrom(matrixStats,rowVars)
 importFrom(methods,as)
