### 2024-03-26 limengbinggz/ddtlcm#11 @limengbinggz
Thank you for pointing this limitation out. We have added a function ``plot.ddt_lcm`` to the package to produce trace plots of model parameters. The users can specify which parameters they are interested in and create a sequence of trace plots of the posterior chain after burn-ins. This function is added in commit [d7621ba](https://github.com/limengbinggz/ddtlcm/commit/d7621bab58d857f58ad59fcf4bf214fed9262588).
