+++ JOSS/references.bib
@@ -86,3 +86,10 @@
   issn = {0361-0918},
   doi = {10.1080/03610918.2012.718840}
 }
+
+@article{talts2018validating,
+  title={Validating Bayesian inference algorithms with simulation-based calibration},
+  author={Talts, Sean and Betancourt, Michael and Simpson, Daniel and Vehtari, Aki and Gelman, Andrew},
+  journal={arXiv preprint arXiv:1804.06788},
+  year={2018}
+}
