+++ inst/REFERENCES.bib
@@ -1,30 +1,0 @@
-@article{knowles2015pitman,
-  title={Pitman-{Yor} diffusion trees for {Bayesian} hierarchical clustering},
-	volume = {37},
-	issn = {1939-3539},
-	doi = {10.1109/TPAMI.2014.2313115},
-	language = {eng},
-	number = {2},
-	journal = {IEEE transactions on pattern analysis and machine intelligence},
-	author = {Knowles, David A. and Ghahramani, Zoubin},
-	month = feb,
-	year = {2015},
-	pmid = {26353241},
-	pages = {271--289}
-}
-
-@article{li2023tree,
-  title={Tree-Regularized Bayesian Latent Class Analysis for Improving Weakly Separated Dietary Pattern Subtyping in Small-Sized Subpopulations},
-  author={Li, Mengbing and Stephenson, Briana and Wu, Zhenke},
-  journal={arXiv preprint arXiv:2306.04700},
-  year={2023}
-}
-
-@article{neal2003density,
-  title={Density modeling and clustering using {Dirichlet} diffusion trees},
-  author={Neal, Radford M},
-  journal={Bayesian Statistics},
-  volume={7},
-  pages={619--629},
-  year={2003}
-}
