### 2024-03-30 404175d
update documentation for CRAN
