### 2024-04-03 689b4d8
closes #15; change vignettes link to follow CRAN rules
