### 2024-05-19 #6220 @Nikoleta-v3
Hello @limengbinggz,

I am very sorry for the long silence and the delay in processing your submission. Life has been hectic the past few weeks. Regardless, apologies for the delay!

I had some time to go over it as well. I have a few questions and comments.

**Regarding the paper:**
- Could you please fix the typos: line 89, repeated `of`; line 94, `burn-ins`.
- Subsections such as `#### Data Loading` are hard to notice. You are using a fourth-level heading. Why not use a second-level heading since they are subsections?

**Regarding the package:**

I can see you have a few tests, but I didn’t find any guidelines on how to run them, which brings me to the next point: documentation. The package is documented, but I believe it can be improved. I could only find the quick start and the workflow example in the [vignettes](https://github.com/limengbinggz/ddtlcm/tree/main/vignettes) from the project's repository.

A few comments about the documentation:

- In the README, examples come after the `Quickstart`. It's common practice to have these sections the other way around.

- Regarding the `ddtlcm-demo`, the link should point the user to the CRAN vignette: https://cran.r-project.org/web/packages/ddtlcm/vignettes/ddtlcm-demo.html instead of the markdown file because it's more readable.

- In the CRAN project, there is also a reference manual, which I believe is very useful for users. This should be mentioned in the README.

**Bullet points**

In the README, you use bullet points in the overview and examples sections. In the overview, you state:

`ddtlcm` works for:
- Multivariate binary responses over pre-specified grouping of items
- The functions' relations in the package `ddtlcm` can be visualized by

Only the first bullet point is relevant to the sentence "`ddtlcm` works for:". So the second bullet point should just be a sentence.

You do this in the examples section too. You point the user to the workflow, but you also state `ddtlcm estimates the tree over classes and class profiles simultaneously`. I don't believe that this sentence should be a bullet point. If you would like to include a picture to showcase the functionality of the package, maybe consider having a "Gallery" section. See this README for example: https://github.com/marcharper/python-ternary.

**Overall**

Please include information on how to run the tests of the package. This is useful in case someone wants to contribute to the package. This way, they can test whether their contribution "broke" something. Please correct some issues with the README and reorganize some sections.

Moreover (this is a recommendation), I suggest you work on the documentation of the package a bit more. Include some more vignettes.
