### 2024-07-01 975b5c8
corrected typos in JOSS paper; corrected README according to JOSS editor comment
