### 2024-07-01 b9556b2
update joss workflow file
