### 2024-07-04 #6220 @Nikoleta-v3
Hello @limengbinggz, thank you for your work 😄

Given the time it took to get back to us, I have to admit that I expected you would have added a few more vignettes  to the documentation of the package.

However, I will leave this as a "future work" for you on your to-do list.

For now, could you please:

- (If necessary) Make a new tagged release of your software and list the version tag of the archived version here. Alternatively, let me know if I should just use release v0.1.2.
- Archive the reviewed software in Zenodo or a similar service (e.g., Figshare, an institutional repository).
- Ensure the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (make sure the list is correct and that people who only made small fixes are not included). You may also add the authors' ORCID IDs.
- List the DOI of the archived version here.
