### 2024-07-10 #6220 @limengbinggz
> Hello @limengbinggz, thank you for your work 😄
>
> Given the time it took to get back to us, I have to admit that I expected you would have added a few more vignettes to the documentation of the package.
>
> However, I will leave this as a "future work" for you on your to-do list.
>
> For now, could you please:
>
- [x] (If necessary) Make a new tagged release of your software and list the version tag of the archived version here. Alternatively, let me know if I should just use release v0.1.2.
I have made a new tagged release of the software. The latest release is [v0.2.1](https://github.com/limengbinggz/ddtlcm/releases/tag/v0.2.1). The archived release is [0.1.1](https://github.com/limengbinggz/ddtlcm/releases/tag/0.1.1).
- [x] Archive the reviewed software in Zenodo or a similar service (e.g., Figshare, an institutional repository).
- [x] Ensure the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (make sure the list is correct and that people who only made small fixes are not included). You may also add the authors' ORCID IDs.
- [x] List the DOI of the archived version here.
DOI: 10.5281/zenodo.12711232

Sorry about the delay as I was not aware of the comment.

* I appreciate your advice on the vignettes. Could you please give me more specific suggestions about what can be added to the vignettes for improvement? Thank you!
