### 2024-07-11 #6220 @Nikoleta-v3
> I appreciate your advice on the vignettes. Could you please give me more specific suggestions about what can be added to the vignettes for improvement? Thank you!

The current vignette is really to accompany your paper. I was thinking of having vignettes for the different functionalities of the package. Are all the functionalities of `ddtlcm` being captured by the current vignette?
