### 2024-07-11 #6220 @limengbinggz
> > I appreciate your advice on the vignettes. Could you please give me more specific suggestions about what can be added to the vignettes for improvement? Thank you!
>
> The current vignette is really to accompany your paper. I was thinking of having vignettes for the different functionalities of the package. Are all the functionalities of `ddtlcm` being captured by the current vignette?

Thanks! I do think the current vignettes is to demonstrate all the functionalities of the package, basically data simulation, model estimation and inference, and visualization.
