### 2024-07-11 #6220 @limengbinggz
> @limengbinggz can you check if you can find DOIs for the above publications? ⬆️

I have updated the DOI for randomLCM. The other two references do not have DOIs so I am leaving them blank. Thanks!
