### 2024-07-15 #6220 @limengbinggz
> https://github.com/limengbinggz/ddtlcm

> 👋 @limengbinggz - Thanks for your work on this. We just need to clear one more thing before I can accept this for publication:
>
> * [x]  The license listed in your Zenodo archive should match the license listed for your software.  Please edit the metadata of your Zenodo archive to make this so.  No need to conduct a new release.
>
> Let me know when this as done and we will move forward. Thanks!

Thank you for the nudge! I have corrected my license on Zenodo.
