### 2024-07-15 #6220 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
